refactor(llm): move dialect-neutral helpers into the codec layer
parse_error_body and extract_system_prompt are pure wire-translation
helpers consumed by the codecs (and the Codec::decode_error default);
they move from providers::common into codec/mod.rs so the codec layer
no longer imports from the transport-side providers module.
providers::common shrinks to route helpers (catalog_model/api_model_id),
attachment file loading, and the frozen re-export shims.
